+++ src/com.bank/domain/user/Menager.java
@@ -1,23 +1,24 @@
 package com.bank.domain.user;
 
+import java.math.BigDecimal;
 import java.time.LocalDate;
 
 public non-sealed class  Menager extends User {
 
-    double salary;
+    BigDecimal salary;
 
 
-    public Menager(String name, String cpf, String password, LocalDate birthDate, double salary){
+    public Menager(String name, String cpf, String password, LocalDate birthDate, BigDecimal salary){
         super(name, cpf, password, birthDate);
         super.setAgency(null);
         this.salary = salary;
     }
 
-    public double getSalary() {
+    public BigDecimal getSalary() {
         return salary;
     }
 
-    public void setSalary(double salary) {
+    public void setSalary(BigDecimal salary) {
         this.salary = salary;
     }
 }
